+++ src/vanilla/Model/MethodPy.cs
@@ -447,7 +447,7 @@ namespace AutoRest.Python.Model
 
             if (HasResponseHeader)
             {
-                if (code != null)
+                if (code == null)
                 {
                     headersType = this.ReturnType.Headers;
                 }
@@ -651,8 +651,11 @@ namespace AutoRest.Python.Model
                 {
                     builder.AppendLine("{0} = None",
                             transformation.OutputParameter.Name);
-                    builder.AppendLine("if {0}:", BuildNullCheckExpression(transformation))
-                       .Indent();
+                    if (!transformation.ParameterMappings.IsNullOrEmpty())
+                    {
+                        builder.AppendLine("if {0}:", BuildNullCheckExpression(transformation))
+                        .Indent();
+                    }
                     foreach (var mapping in transformation.ParameterMappings)
                     {
                         builder.AppendLine(mapping.CreateCode(transformation.OutputParameter));