+++ test/vanilla/Expected/AcceptanceTests/Validation/auto_rest_validation_test/auto_rest_validation_test.py
@@ -32,8 +32,6 @@ class AutoRestValidationTestConfiguration(Configuration):
 
         if subscription_id is None:
             raise ValueError("Parameter 'subscription_id' must not be None.")
-        if not isinstance(subscription_id, str):
-            raise TypeError("Parameter 'subscription_id' must be str.")
         if not base_url:
             base_url = 'http://localhost'