+++ ChangeLog.md
@@ -1,5 +1,13 @@
+### 2017-10-10 - 2.0.16
+- Improve polymorphic discriminator documentation (#17)
+- Add deprecated support (#16)
+- Fix invalid headers test (#15)
+- Fix invalid python code in some scenario (#14)
+- Stop checking str type for client parameter (#12)
+- Add client-side-validation to Autorest.Python (#11)
+
 ### 2017-09-27 - 2.0.14
-- Improve documentation cross-reference
+- Improve documentation cross-reference (#9)
 
 ### 2017-09-26 - 2.0.13
-- Remove constraint on array type, if array is used in the URL
+- Remove constraint on array type, if array is used in the URL (#8)